pyproject.toml
[project]
name = "kkbox2applemusic"
version = "0.1.0"
description = "將 KKBOX 播放清單轉換為 Apple Music 格式"
readme = "README.md"
authors = [
    { name = "Donald Lee", email = "donald.lee@macrowell.com.tw" }
]
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.7",
    "httpx>=0.28.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.24.1",
]

[project.scripts]
kkbox2applemusic = "kkbox2applemusic:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
